refactor: 페이지 및 현재 입력된 키워드 관련 변수 및 함수 분리
