+++ src/data/useKeyword.ts
@@ -1,0 +1,9 @@
+let recentKeyword: string;
+
+export function getRecentKeyword() {
+  return recentKeyword;
+}
+
+export function setRecentKeyword(keyword: string) {
+  recentKeyword = keyword;
+}
